+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 Change Log
 ==========
 
-2019-10-18
-----------
+2019-10-25 - 1.0.2
+------------------
+
+ * Added controller argument resolvers for UUID and ExternalIdentity
+ * Added includes() to ApiController and RequestArgumentHelper service
+
+2019-10-22 - 1.0.1
+------------------
+
+ * Fixed bug with alias of perPage (was per_page)
+
+2019-10-18 - 1.0.0
+------------------
 
 Initial commit.